[FAB-11553] Split LSCC validation
For the sake of readability, the validation logic for lscc is split into 
its own compilation unit.
